pyproject.toml
[tool.black]
target-version = ['py311']
line-length = 121
force-exclude = '''
migrations
'''

[tool.ruff]
exclude = ["migrations"]
line-length = 121
ignore = [
    # doesn't like Optional
    # https://github.com/charliermarsh/ruff/issues/4858
    "UP007",
    # explicit raise from might be a bit too verbose, disable for now
    "B904",
    # asserts are useful in both tests and type checker persuasion
    "S101",
    # it is annoying to annotate django Meta model properties as CalssVar for no reason
    "RUF012",
    # treats a link with the word "password" in it as a hardcoded password
    "S105",
]
select = [
    # pyflakes
    "F",
    # pycodestyle
    "E",
    "W",
    # isort
    "I",
    # flake8-bugbear
    "B",
    # flake8-django
    "DJ",
    # flake8-naming
    "N",
    # pyupgrade
    "UP",
    # flake8-bandit
    "S",
    # flake8-comprehensions
    "C4",
    # flake8-logging-format
    "G",
    # flake8-simplify
    "SIM",
    # flake8-use-pathlib
    "PTH",
    # ruff
    "RUF",
]
fixable = [
    "I",
]

[tool.ruff.isort]
combine-as-imports = true
lines-between-types = 1

# with 'mypy' and 'django-stubs[compatible-mypy]' installed
# [tool.mypy]
# plugins = ["mypy_django_plugin.main"]
#
# [tool.django-stubs]
# django_settings_module = "central_command.settings"
#
# [[tool.mypy.overrides]]
# module = [
#     "django_email_verification.*",
#     "rest_framework.*",
#     "knox.*",
# ]
# ignore_missing_imports = true

[tool.poetry]
name = "central-command"
version = "0.1.0"
description = "All-in-one backend application for Unitystation"
authors = ["Andrés Riquelme <andresfranco.rs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^3.2.12"
djangorestframework = "^3.12.1"
psycopg2-binary = "2.9.9"
django-rest-knox = "^4.1.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.2"
whitenoise = "^6.2.0"
django-post-office = "^3.8.0"
drf-spectacular = "^0.27.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
ruff = "^0.1.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
